requirements.txt
streamlit

tqdm
streamlit_code_editor
pandas
openpyxl
pyyaml